Small refactor to keep classes in separate files
