+++ engine/launcher/src/main/scala/org/enso/launcher/releases/Asset.scala
@@ -1,0 +1,14 @@
+package org.enso.launcher.releases
+
+import java.nio.file.Path
+
+import org.enso.cli.TaskProgress
+
+/**
+  * Represents a downloadable release asset.
+  */
+trait Asset {
+  def fileName:               String
+  def downloadTo(path: Path): TaskProgress[Unit]
+  def fetchAsText():          TaskProgress[String]
+}
