+++ engine/launcher/src/main/scala/org/enso/launcher/releases/ReleaseProviderException.scala
@@ -1,0 +1,11 @@
+package org.enso.launcher.releases
+
+case class ReleaseProviderException(message: String, cause: Throwable = null)
+    extends RuntimeException(message, cause) {
+
+  /**
+    * @inheritdoc
+    */
+  override def toString: String =
+    s"A problem occurred when trying to find the release: $message"
+}
